requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
discord.py==2.5.2
distlib==0.3.8
Django==5.0.6
django-debug-toolbar==4.3.0
filelock==3.14.0
frozenlist==1.5.0
idna==3.10
multidict==6.3.2
mysqlclient==2.2.4
platformdirs==4.2.1
propcache==0.3.1
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.26.1
yarl==1.18.3
discord.py
twitchAPI
python-dotenv
requests